CB-1375 blueprintUtil endpoint was not in effect
