+++ core-api/src/main/java/com/sequenceiq/cloudbreak/api/endpoint/v4/blueprint/BlueprintUtilV4Endpoint.java
@@ -13,6 +13,7 @@ import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 
 import com.sequenceiq.cloudbreak.api.endpoint.v4.blueprint.responses.GeneratedCmTemplateV4Response;
+import com.sequenceiq.cloudbreak.api.endpoint.v4.blueprint.responses.RecommendationV4Response;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.blueprint.responses.ServiceDependencyMatrixV4Response;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.blueprint.responses.SupportedVersionsV4Response;
 import com.sequenceiq.cloudbreak.doc.ControllerDescription;
@@ -22,9 +23,9 @@ import com.sequenceiq.cloudbreak.doc.OperationDescriptions.ConnectorOpDescriptio
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 
-@Path("/v4/{workspaceId}/blueprints-util")
+@Path("/v4/{workspaceId}/blueprints_util")
 @Consumes(MediaType.APPLICATION_JSON)
-@Api(value = "/v4/{workspaceId}/blueprints-util", description = ControllerDescription.BLUEPRINT_V4_DESCRIPTION, protocols = "http,https")
+@Api(value = "/v4/{workspaceId}/blueprints_util", description = ControllerDescription.BLUEPRINT_V4_DESCRIPTION, protocols = "http,https")
 public interface BlueprintUtilV4Endpoint {
 
     @GET
@@ -50,4 +51,13 @@ public interface BlueprintUtilV4Endpoint {
     GeneratedCmTemplateV4Response getGeneratedTemplate(@PathParam("workspaceId") Long workspaceId,
         @QueryParam("services") Set<String> services, @QueryParam("platform") String platform);
 
+    @GET
+    @Path("recommendation")
+    @Produces(MediaType.APPLICATION_JSON)
+    @ApiOperation(value = ConnectorOpDescription.GET_RECOMMENDATION, produces = JSON, notes = Notes.CONNECTOR_NOTES,
+            nickname = "createRecommendationForWorkspace")
+    RecommendationV4Response createRecommendation(@PathParam("workspaceId") Long workspaceId,
+            @QueryParam("blueprintName") String blueprintName, @QueryParam("credentialName") String credentialName,
+            @QueryParam("region") String region, @QueryParam("platformVariant") String platformVariant,
+            @QueryParam("availabilityZone") String availabilityZone);
 }
