+++ core-api/src/main/java/com/sequenceiq/cloudbreak/client/CloudbreakClient.java
@@ -30,6 +30,7 @@ import org.slf4j.LoggerFactory;
 import com.sequenceiq.cloudbreak.api.CoreApi;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.audits.AuditEventV4Endpoint;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.autoscales.AutoscaleV4Endpoint;
+import com.sequenceiq.cloudbreak.api.endpoint.v4.blueprint.BlueprintUtilV4Endpoint;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.blueprint.BlueprintV4Endpoint;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.clustertemplate.ClusterTemplateV4Endpoint;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.connector.ConnectorV4Endpoint;
@@ -66,6 +67,7 @@ public class CloudbreakClient {
             AuditEventV4Endpoint.class,
             AutoscaleV4Endpoint.class,
             BlueprintV4Endpoint.class,
+            BlueprintUtilV4Endpoint.class,
             DatabaseV4Endpoint.class,
             EnvironmentV4Endpoint.class,
             EventV4Endpoint.class,
@@ -124,6 +126,10 @@ public class CloudbreakClient {
 
     public BlueprintV4Endpoint blueprintV4Endpoint() {
         return getEndpoint(BlueprintV4Endpoint.class);
+    }
+
+    public BlueprintUtilV4Endpoint blueprintUtilV4Endpoint() {
+        return getEndpoint(BlueprintUtilV4Endpoint.class);
     }
 
     public EnvironmentV4Endpoint environmentV4Endpoint() {
