+++ core-api/src/main/java/com/sequenceiq/cloudbreak/client/CloudbreakIdentityClient.java
@@ -28,6 +28,7 @@ import org.slf4j.LoggerFactory;
 import com.sequenceiq.cloudbreak.api.CoreApi;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.audits.AuditEventV4Endpoint;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.autoscales.AutoscaleV4Endpoint;
+import com.sequenceiq.cloudbreak.api.endpoint.v4.blueprint.BlueprintUtilV4Endpoint;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.blueprint.BlueprintV4Endpoint;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.clustertemplate.ClusterTemplateV4Endpoint;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.connector.ConnectorV4Endpoint;
@@ -58,6 +59,7 @@ public class CloudbreakIdentityClient {
             AuditEventV4Endpoint.class,
             AutoscaleV4Endpoint.class,
             BlueprintV4Endpoint.class,
+            BlueprintUtilV4Endpoint.class,
             EventV4Endpoint.class,
             ClusterTemplateV4Endpoint.class,
             CredentialV4Endpoint.class,
@@ -135,6 +137,10 @@ public class CloudbreakIdentityClient {
 
     public BlueprintV4Endpoint blueprintV4Endpoint() {
         return getEndpoint(BlueprintV4Endpoint.class);
+    }
+
+    public BlueprintUtilV4Endpoint blueprintUtilV4Endpoint() {
+        return getEndpoint(BlueprintUtilV4Endpoint.class);
     }
 
     public EventV4Endpoint eventV3Endpoint() {
