+++ core-api/src/main/java/com/sequenceiq/cloudbreak/client/CloudbreakUserCrnClient.java
@@ -17,6 +17,7 @@ import org.slf4j.LoggerFactory;
 import com.sequenceiq.cloudbreak.api.CoreApi;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.audits.AuditEventV4Endpoint;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.autoscales.AutoscaleV4Endpoint;
+import com.sequenceiq.cloudbreak.api.endpoint.v4.blueprint.BlueprintUtilV4Endpoint;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.blueprint.BlueprintV4Endpoint;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.clustertemplate.ClusterTemplateV4Endpoint;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.connector.ConnectorV4Endpoint;
@@ -85,6 +86,10 @@ public class CloudbreakUserCrnClient {
 
         public BlueprintV4Endpoint blueprintV4Endpoint() {
             return getEndpoint(BlueprintV4Endpoint.class);
+        }
+
+        public BlueprintUtilV4Endpoint blueprintUtilV4Endpoint() {
+            return getEndpoint(BlueprintUtilV4Endpoint.class);
         }
 
         public EnvironmentV4Endpoint environmentV4Endpoint() {
